add BathroomCalculator and integrate daily percentage calculation in CalculationResponseMapper

+++ src/main/java/com/tronoremunerado/calculator/domain/KingCalculateResponse.java
@@ -10,6 +10,6 @@ public record KingCalculateResponse(
         BigDecimal dailyEarnings,
         BigDecimal monthlyEarnings,
         BigDecimal yearlyEarnings,
-        BigDecimal dailyPercentageOfShift
+        double dailyPercentageOfShift
 ) {}
